pyproject.toml
[project]
name = "varphi-devkit"
version = "1.5.0"
description = "A Python framework for creating compilers that target the Varphi language"
authors = [
    {name = "Hassan El-Sheikha",email = "hmelsheikha@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
keywords = ["compiler", "turing-machine", "dsl", "antlr", "parser"]
license = {text = "BSD-3-Clause"}
homepage = "https://github.com/varphi-lang/varphi-devkit"
repository = "https://github.com/varphi-lang/varphi-devkit"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Compilers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "antlr4-python3-runtime (>=4.13.2,<5.0.0)",
]

[tool.poetry]
packages = [{include = "varphi_devkit", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pylint = "^3.3.7"
antlr4-tools = "^0.2.2"
python-semantic-release = "^10.2.0"
build = "^1.3.0"
twine = "^6.1.0"
black = "^25.12.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[semantic_release]
assets = []
build_command_env = []
commit_message = "{version}\n\nAutomatically generated by python-semantic-release"
commit_parser = "conventional"
logging_use_named_masks = false
major_on_zero = true
allow_zero_version = false
repo_dir = "."
no_git_verify = false
tag_format = "v{version}"
version_toml = ["pyproject.toml:project.version"]

[semantic_release.branches.main]
match = "(main|master)"
prerelease_token = "rc"
prerelease = false

[semantic_release.changelog]
exclude_commit_patterns = []
mode = "update"
insertion_flag = "<!-- version list -->"
template_dir = "templates"

[semantic_release.changelog.default_templates]
changelog_file = "CHANGELOG.md"
output_format = "md"
mask_initial_release = true

[semantic_release.changelog.environment]
block_start_string = "{%"
block_end_string = "%}"
variable_start_string = "{{"
variable_end_string = "}}"
comment_start_string = "{#"
comment_end_string = "#}"
trim_blocks = false
lstrip_blocks = false
newline_sequence = "\n"
keep_trailing_newline = false
extensions = []
autoescape = false

[semantic_release.commit_author]
env = "GIT_COMMIT_AUTHOR"
default = "semantic-release <semantic-release>"

[semantic_release.commit_parser_options]
minor_tags = ["feat"]
patch_tags = ["fix", "perf"]
other_allowed_tags = ["build", "chore", "ci", "docs", "style", "refactor", "test"]
allowed_tags = ["feat", "fix", "perf", "build", "chore", "ci", "docs", "style", "refactor", "test"]
default_bump_level = 0
parse_squash_commits = true
ignore_merge_commits = true

[semantic_release.remote]
name = "origin"
type = "github"
ignore_token_for_push = false
insecure = false

[semantic_release.publish]
dist_glob_patterns = ["dist/*"]
upload_to_vcs_release = true

